requirements.txt
streamlit
spacy
spacy-transformers
torch
pdfminer.six
pandas
plotly